fixed functionality of comments and likes

+++ api/api.js
@@ -5,27 +5,31 @@ export class Api {
     static async addLike(req, res) {
         /////////////////////////////////////////////////// fix
         const user_id = await getUserIdByName(req.session.username);
+        if (user_id === "ERROR") {
+            req.session.name = undefined;
+            return res.redirect('/signin');
+        }
 
         const { comment_id } = req.body;
 
         // check if is already liked
 
         const isLiked = (await getLike(user_id, comment_id)) !== undefined;
-        console.log("isLiked:", isLiked);
+        //console.log("isLiked:", isLiked);
 
         let action;
         // if yes delete like
         if (isLiked) {
             // find reaction_id TODO
             const reaction_id = await getReactionId(user_id, comment_id);
-            console.log("reaction_id", reaction_id);
+            //console.log("reaction_id", reaction_id);
 
             action = "deleted";
             await deleteReaction(reaction_id);
         }
         else {
             // else add like
-            console.log("comment id: ", comment_id, " user id: ", user_id);
+            //console.log("comment id: ", comment_id, " user id: ", user_id);
             await addReaction("like", comment_id, user_id);
             action = "added";
         }
@@ -39,6 +43,39 @@ export class Api {
         });
     }
 
+    static async removeLike(req, res) {
+        const user_id = await getUserIdByName(req.session.username);
+        if (user_id === "ERROR") {
+            req.session.name = undefined;
+            return res.redirect('/signin');
+        }
+
+        const { comment_id } = req.body;
+
+        // check if is already liked
+
+        const isLiked = (await getLike(user_id, comment_id)) !== undefined;
+        //console.log("isLiked:", isLiked);
+
+        let action = "";
+        // if yes delete like
+        if (isLiked) {
+            // find reaction_id TODO
+            const reaction_id = await getReactionId(user_id, comment_id);
+            //console.log("reaction_id", reaction_id);
+
+            action = "like-deleted";
+            await deleteReaction(reaction_id);
+        }
+        let likescount = await getLikes(comment_id);
+
+        return res.json({
+            likescount: likescount,
+            action: action
+        });
+    }
+
+
     static async addComment(req, res) {
         const { post_id, content } = req.body;
         const user_id = await getUserIdByName(req.session.username);
@@ -51,8 +88,10 @@ export class Api {
 
         // add comment
         await addComment(content, post_id, user_id)
+        //console.log("comment added: ", content, " post id: ", post_id, " user id: ", user_id);
 
         res.redirect(`/beach/${post_id}`);
+
     }
 
     static test(req, res) {
